+++ scripts/front-end/components/Dossier/EnteteDossier.svelte
@@ -8,8 +8,6 @@
     /** @type {DossierComplet} */
     export let dossier
 
-    const {number_demarches_simplifiées: numdos} = dossier
-
     let phase = dossier.évènementsPhase[0].phase
 
 </script>
@@ -77,25 +75,11 @@
         <!--
         <div>
             <span class="fr-icon-scales-3-fill" aria-hidden="true"></span>
-            Contentiex
-            {dossier.rattaché_au_régime_ae}
+            Contentieux
         </div>
         -->
     </section>
 </header>
-
-<!--
-<nav class="dossier-nav fr-mb-4w">
-    <ul class="fr-btns-group fr-btns-group--inline fr-btns-group--sm fr-mb-2w">
-        <li> 
-            <a class="fr-btn fr-btn--secondary fr-my-0" target="_blank" href={`https://www.demarches-simplifiees.fr/procedures/88444/dossiers/${numdos}`}>Dossier sur Démarches Simplifiées</a>
-        </li>
-        <li>
-            
-        </li>
-    </ul>
-</nav>
--->
 
 <style lang="scss">
     header{
